[Attention] Adapt ScaledDotProductAttention to recent frontend changes
Note: This is a minimally invasive adaptation to make the attention
operator handle the batch/head dimension which is not squeezed out of
the model anymore. A proper refactoring should be done later.

Includes a bugfix to the code generation for thresholds embedded into
the attention operator: Apprently comp::less is correct after all...